backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scapy==2.5.0
werkzeug==3.0.1
PyJWT==2.8.0